+++ lib/main.dart
@@ -2,72 +2,48 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:study_riverpod/test_provider.dart';
 
 void main() {
-  // Providerをアプリで使用するためにProviderScopeで囲む
-  runApp(
-    const ProviderScope(
-      child: MyApp(),
-    ),
-  );
+  runApp(const ProviderScope(child: MyApp()));
 }
 
-class MyApp extends ConsumerWidget {
+class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
+  // This widget is the root of your application.
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final title = ref.watch(
-      titleProvider,
-    );
-
-    final label = ref.watch(
-      labelProvider(
-        str1: 'テスト',
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
       ),
+      home: const MyHomePage(title: 'Flutter Demo Home Page'),
     );
+  }
+}
 
-    final name = ref.watch(
-      nameProvider,
-    );
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({required this.title, super.key});
 
-    final count = ref.watch(
-      countProvider,
-    );
+  final String title;
 
-    // FutureやStreamのProviderはwhenを使うこと状態に応じた処理が可能
-    return name.when(
-      loading: CircularProgressIndicator.adaptive,
-      error: (err, stack) => Center(child: Text('error: $err')),
-      data: (name) {
-        return count.when(
-          loading: CircularProgressIndicator.adaptive,
-          error: (err, stack) => Center(child: Text('error: $err')),
-          // 複数の状態を待ちたい場合はwhenをネストするしか無い？
-          data: (count) {
-            return MaterialApp(
-              title: 'Flutter Demo',
-              theme: ThemeData(
-                colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-                useMaterial3: true,
-              ),
-              home: Scaffold(
-                appBar: AppBar(title: Text(title)),
-                body: Center(
-                  child: Column(
-                    children: [
-                      Text(label),
-                      Text(name),
-                      Text(count),
-                    ],
-                  ),
-                ),
-              ),
-            );
-          },
-        );
-      },
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: Text(widget.title),
+      ),
+      body: const Center(
+        child: Column(),
+      ),
     );
   }
 }
